# docs/openapi/zerg.v1.json
{
  "components": {
    "responses": {
      "BadRequestUuidResponse": {
        "content": {
          "application/json": {
            "example": {
              "code": 1002,
              "details": null,
              "error": "INVALID_UUID",
              "message": "Invalid UUID format"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Bad Request - Invalid UUID"
      },
      "BadRequestValidationResponse": {
        "content": {
          "application/json": {
            "example": {
              "code": 1001,
              "details": {
                "title": [
                  {
                    "code": "length",
                    "message": "length is less than 3",
                    "params": {
                      "min": 3,
                      "value": "ab"
                    }
                  }
                ]
              },
              "error": "VALIDATION_ERROR",
              "message": "Request validation failed"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Bad Request - Validation Error"
      },
      "ConflictResponse": {
        "content": {
          "application/json": {
            "example": {
              "code": 1008,
              "details": null,
              "error": "CONFLICT",
              "message": "Resource already exists"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Conflict - Resource already exists"
      },
      "InternalServerErrorResponse": {
        "content": {
          "application/json": {
            "example": {
              "code": 1005,
              "details": null,
              "error": "INTERNAL_ERROR",
              "message": "An internal server error occurred"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Internal Server Error"
      },
      "NotFoundResponse": {
        "content": {
          "application/json": {
            "example": {
              "code": 1004,
              "details": null,
              "error": "NOT_FOUND",
              "message": "Resource not found"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Resource not found"
      }
    },
    "schemas": {
      "CloudProvider": {
        "description": "Supported cloud providers",
        "enum": [
          "aws",
          "gcp",
          "azure",
          "local"
        ],
        "type": "string"
      },
...
      "CreateProject": {
        "description": "DTO for creating a new project",
        "properties": {
          "budget_limit": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "cloud_provider": {
            "$ref": "#/components/schemas/CloudProvider"
          },
          "description": {
            "type": "string"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "user_id",
          "cloud_provider",
          "region"
        ],
        "type": "object"
      },
...
      "Environment": {
        "description": "Environment type for the project",
        "enum": [
          "development",
          "staging",
          "production"
        ],
        "type": "string"
      },
      "ErrorResponse": {
        "description": "Standard error response structure.\n\nThis structure is returned for all error responses, providing consistent\nerror information to clients including\n- `code`: Integer error code for logging/monitoring (e.g., 1008)\n- `error`: Machine-readable error identifier (e.g., \"CONFLICT\")\n- `message`: Human-readable error message\n- `details`: Optional additional error details (e.g., validation errors)\n\n# JSON Example\n\n```json\n{\n  \"code\": 1008,\n  \"error\": \"CONFLICT\",\n  \"message\": \"Resource already exists\",\n  \"details\": null\n}\n```",
        "properties": {
          "code": {
            "description": "Integer error code for logging and monitoring",
            "format": "int32",
            "type": "integer"
          },
          "details": {
            "description": "Optional structured error details (e.g., validation field errors)"
          },
          "error": {
            "description": "Machine-readable error identifier for programmatic handling",
            "type": "string"
          },
          "message": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "Project": {
        "description": "Project entity - represents a cloud project",
        "properties": {
          "budget_limit": {
            "description": "Optional monthly budget limit in USD",
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "cloud_provider": {
            "$ref": "#/components/schemas/CloudProvider",
            "description": "Cloud provider (AWS, GCP, Azure)"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "description": "Project description",
            "type": "string"
          },
          "enabled": {
            "description": "Whether the project is enabled",
            "type": "boolean"
          },
          "environment": {
            "$ref": "#/components/schemas/Environment",
            "description": "Environment type"
          },
          "id": {
            "description": "Unique identifier",
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "description": "Project name (must be unique per user)",
            "type": "string"
          },
          "region": {
            "description": "Deployment region (e.g., us-east-1, europe-west1)",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Current status"
          },
          "tags": {
            "description": "Resource tags for organization",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array"
          },
          "updated_at": {
            "description": "Last update timestamp",
            "format": "date-time",
            "type": "string"
          },
          "user_id": {
            "description": "Owner of the project",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "user_id",
          "description",
          "cloud_provider",
          "region",
          "environment",
          "status",
          "tags",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "ProjectStatus": {
        "description": "Project deployment status",
        "enum": [
          "provisioning",
          "active",
          "suspended",
          "deleting",
          "archived"
        ],
        "type": "string"
      },
...
      "Tag": {
        "description": "Key-value tag for project organization",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "type": "object"
      },
...
      "UpdateProject": {
        "description": "DTO for updating an existing project",
        "properties": {
          "budget_limit": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "environment": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Environment"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProjectStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "in": "cookie",
        "name": "zerg_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API for managing tasks, projects, cloud resources, and users",
    "license": {
      "name": ""
    },
    "title": "Zerg API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "list_projects",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "cloud_provider",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/CloudProvider"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Environment"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ProjectStatus"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "enabled",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of projects"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "List projects with optional filters",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project created successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestValidationResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "delete": {
        "operationId": "delete_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestUuidResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "get_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project found"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestUuidResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Get a project by ID",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "operationId": "update_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project updated successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestValidationResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Update a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/activate": {
      "post": {
        "operationId": "activate_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project activated successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestUuidResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Activate a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/archive": {
      "post": {
        "operationId": "archive_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project archived successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestUuidResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Archive a project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}/suspend": {
      "post": {
        "operationId": "suspend_project",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Project suspended successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestUuidResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ],
        "summary": "Suspend a project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "API base path",
      "url": "/api"
    }
  ],
...
}